MLOps-DAG/requirements.txt
pandas
scikit-learn
tensorflow
numpy
mlflow
apache-airflow